requirements.txt
asgiref==3.8.1
boto==2.49.0
boto3==1.37.24
botocore==1.37.24
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
jmespath==1.0.1
pillow==11.1.0
psycopg==3.2.6
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0